// frontend/sdk/models/PlaylistSortBy.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sopranino API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const PlaylistSortBy = {
    Default: 'default',
    Name: 'name',
    Tracks: 'tracks'
} as const;
export type PlaylistSortBy = typeof PlaylistSortBy[keyof typeof PlaylistSortBy];


export function instanceOfPlaylistSortBy(value: any): boolean {
    for (const key in PlaylistSortBy) {
        if (Object.prototype.hasOwnProperty.call(PlaylistSortBy, key)) {
            if (PlaylistSortBy[key as keyof typeof PlaylistSortBy] === value) {
                return true;
            }
        }
    }
    return false;
}

export function PlaylistSortByFromJSON(json: any): PlaylistSortBy {
    return PlaylistSortByFromJSONTyped(json, false);
}

export function PlaylistSortByFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlaylistSortBy {
    return json as PlaylistSortBy;
}

export function PlaylistSortByToJSON(value?: PlaylistSortBy | null): any {
    return value as any;
}

export function PlaylistSortByToJSONTyped(value: any, ignoreDiscriminator: boolean): PlaylistSortBy {
    return value as PlaylistSortBy;
}